+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Nothing (I tend to only add finished Stuff to the Changelog).
 
 
 6.10.24: (Not released yet)
+[FIXED] The Fluid Rendering Height Fix from 6.10.19 was slightly broken for other Mods. Now it works again.
 [FIXED] Some Furnace Recipes being missing due to me accidentially removing most of my own Furnace Recipes in the previous Versions change. XD
 [CHANGED] Smelting any Material based thing other than Dusts and Scraps in a Furnace will give EXP depending on the Tool Quality of the Material + 1.
 [CHANGED] IC2-Exp "Pahoehoe Lava" will now give slightly more outputs when centrifuged, but also uses more Power.